# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: DEAN Orchestration API
  description: |
    The DEAN (Distributed Evolutionary Agent Network) API provides comprehensive endpoints for managing
    evolutionary optimization of code through autonomous agents. This API enables creation and management
    of agents, evolution trials, pattern discovery, and system monitoring.
  version: 1.0.0
  contact:
    name: DEAN Support Team
    email: support@dean-orchestration.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.dean-orchestration.com
    description: Production server
  - url: https://staging.dean-orchestration.com
    description: Staging server
  - url: http://localhost:8082
    description: Local development
security:
  - bearerAuth: []
...
paths:
  /auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticate user and receive access tokens
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            examples:
              admin:
                value:
                  username: admin
                  password: admin123
      responses:
        '200':
          description: Successful authentication
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '429':
          $ref: '#/components/responses/RateLimitError'
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: Use refresh token to get new access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
  /api/agents:
    get:
      tags:
        - Agents
      summary: List all agents
      description: Retrieve a list of all agents with optional filtering
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [active, inactive, evolving]
          description: Filter by agent status
        - in: query
          name: language
          schema:
            type: string
            enum: [python, javascript, java]
          description: Filter by programming language
        - in: query
          name: limit
          schema:
            type: integer
            default: 100
            minimum: 1
            maximum: 1000
          description: Maximum number of results
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
            minimum: 0
          description: Number of results to skip
      responses:
        '200':
          description: List of agents
          content:
            application/json:
              schema:
                type: object
                properties:
                  agents:
                    type: array
                    items:
                      $ref: '#/components/schemas/Agent'
                  total:
                    type: integer
                    example: 42
                  limit:
                    type: integer
                    example: 100
                  offset:
                    type: integer
                    example: 0
        '401':
          $ref: '#/components/responses/UnauthorizedError'

    post:
      tags:
        - Agents
      summary: Create new agent
      description: Create a new evolutionary agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAgentRequest'
            examples:
              pythonAgent:
                value:
                  name: search-optimizer
                  language: python
                  capabilities:
                    - search
                    - optimize
                    - analyze
                  initial_fitness: 0.5
      responses:
        '201':
          description: Agent created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
  /api/evolution/trials:
    get:
      tags:
        - Evolution
      summary: List evolution trials
      description: Get all evolution trials with optional filtering
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [running, completed, failed, cancelled]
          description: Filter by trial status
        - in: query
          name: start_date
          schema:
            type: string
            format: date-time
          description: Filter trials started after this date
        - in: query
          name: end_date
          schema:
            type: string
            format: date-time
          description: Filter trials started before this date
      responses:
        '200':
          description: List of evolution trials
          content:
            application/json:
              schema:
                type: object
                properties:
                  trials:
                    type: array
                    items:
                      $ref: '#/components/schemas/EvolutionTrial'
                  total:
                    type: integer
  /api/evolution/trials/{trialId}:
    get:
      tags:
        - Evolution
      summary: Get trial details
      description: Retrieve detailed information about a specific evolution trial
      parameters:
        - in: path
          name: trialId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Trial details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvolutionTrialDetail'
        '404':
          $ref: '#/components/responses/NotFoundError'
...
  /api/patterns:
    get:
      tags:
        - Patterns
      summary: List discovered patterns
      description: Retrieve patterns discovered during evolution trials
      parameters:
        - in: query
          name: min_confidence
          schema:
            type: number
            format: float
            minimum: 0
            maximum: 1
            default: 0.7
          description: Minimum confidence score
        - in: query
          name: category
          schema:
            type: string
            enum: [optimization, refactoring, security, performance]
          description: Pattern category
        - in: query
          name: language
          schema:
            type: string
            enum: [python, javascript, java]
          description: Programming language
      responses:
        '200':
          description: List of patterns
          content:
            application/json:
              schema:
                type: object
                properties:
                  patterns:
                    type: array
                    items:
                      $ref: '#/components/schemas/Pattern'
                  total:
                    type: integer
...
  /ws:
    get:
      tags:
        - WebSocket
      summary: WebSocket connection
      description: |
        Establish WebSocket connection for real-time updates.
        
        **WebSocket Events:**
        
        Sent by server:
        - `evolution.progress`: Evolution trial progress updates
        - `pattern.discovered`: New pattern discovered
        - `agent.status`: Agent status changes
        - `system.metrics`: Real-time system metrics
        
        Sent by client:
        - `subscribe`: Subscribe to specific event types
        - `unsubscribe`: Unsubscribe from event types
        - `ping`: Keep connection alive
      security:
        - bearerAuth: []
      responses:
        '101':
          description: Switching Protocols - WebSocket connection established
        '401':
          $ref: '#/components/responses/UnauthorizedError'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 50
          example: admin
        password:
          type: string
          minLength: 8
          format: password
          example: admin123
    RefreshRequest:
      type: object
      required:
        - refresh_token
      properties:
        refresh_token:
          type: string
          format: jwt
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
          format: jwt
          description: JWT access token
        refresh_token:
          type: string
          format: jwt
          description: JWT refresh token
        token_type:
          type: string
          example: bearer
        expires_in:
          type: integer
          description: Token expiry time in seconds
          example: 1800
    Agent:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
        name:
          type: string
          example: search-optimizer
        language:
          type: string
          enum: [python, javascript, java]
          example: python
        status:
          type: string
          enum: [active, inactive, evolving]
          example: active
        fitness_score:
          type: number
          format: float
          minimum: 0
          maximum: 1
          example: 0.85
        capabilities:
          type: array
          items:
            type: string
          example: [search, optimize, analyze]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        evolution_count:
          type: integer
          description: Number of evolution trials participated in
          example: 5
    CreateAgentRequest:
      type: object
      required:
        - name
        - language
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 100
          example: search-optimizer
        language:
          type: string
          enum: [python, javascript, java]
          example: python
        capabilities:
          type: array
          items:
            type: string
          example: [search, optimize, analyze]
        initial_fitness:
          type: number
          format: float
          minimum: 0
          maximum: 1
          default: 0.5
...
    StartEvolutionRequest:
      type: object
      required:
        - repository
        - generations
      properties:
        repository:
          type: string
          description: Repository or project to optimize
          example: my-optimization-project
        generations:
          type: integer
          minimum: 1
          maximum: 100
          description: Number of generations to evolve
          example: 10
        population_size:
          type: integer
          minimum: 10
          maximum: 1000
          default: 50
          description: Size of agent population
        mutation_rate:
          type: number
          format: float
          minimum: 0
          maximum: 1
          default: 0.15
          description: Probability of mutation
        crossover_rate:
          type: number
          format: float
          minimum: 0
          maximum: 1
          default: 0.7
          description: Probability of crossover
        fitness_threshold:
          type: number
          format: float
          minimum: 0
          maximum: 1
          default: 0.9
          description: Target fitness score
        selection_method:
          type: string
          enum: [tournament, roulette, rank]
          default: tournament
          description: Selection algorithm
    EvolutionTrial:
      type: object
      properties:
        id:
          type: string
          format: uuid
        repository:
          type: string
        status:
          type: string
          enum: [running, completed, failed, cancelled]
        current_generation:
          type: integer
        total_generations:
          type: integer
        best_fitness:
          type: number
          format: float
        average_fitness:
          type: number
          format: float
        patterns_discovered:
          type: integer
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
    EvolutionTrialDetail:
      allOf:
        - $ref: '#/components/schemas/EvolutionTrial'
        - type: object
          properties:
            generations:
              type: array
              items:
                type: object
                properties:
                  generation:
                    type: integer
                  best_fitness:
                    type: number
                  average_fitness:
                    type: number
                  patterns_found:
                    type: integer
            agents:
              type: array
              items:
                $ref: '#/components/schemas/Agent'
            parameters:
              $ref: '#/components/schemas/StartEvolutionRequest'
    Pattern:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: Async Handler Optimization
        category:
          type: string
          enum: [optimization, refactoring, security, performance]
        confidence:
          type: number
          format: float
          minimum: 0
          maximum: 1
          example: 0.92
        impact_score:
          type: number
          format: float
          minimum: 0
          maximum: 1
          example: 0.85
        description:
          type: string
          example: Convert synchronous operations to async for improved performance
        language:
          type: string
          enum: [python, javascript, java]
        discovered_at:
          type: string
          format: date-time
        usage_count:
          type: integer
          description: Number of times pattern has been applied
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
        code:
          type: string
          description: Error code
        details:
          type: object
          description: Additional error details
  responses:
    UnauthorizedError:
      description: Authentication information is missing or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Unauthorized
            code: AUTH_001
            details:
              message: Invalid or expired token
    BadRequestError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Bad Request
            code: VAL_001
            details:
              field: generations
              message: Value must be between 1 and 100
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Not Found
            code: NOT_001
            details:
              resource: agent
              id: 123e4567-e89b-12d3-a456-426614174000
    RateLimitError:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Too Many Requests
            code: RATE_001
            details:
              limit: 5
              window: 60
              retry_after: 45